Refactor of the parlay enging to make adding executors simpler
