+++ pkg/parlay/parser.go
@@ -2,130 +2,15 @@ package parlay
 
 import (
 	"fmt"
-	"os"
-	"os/exec"
-	"strings"
 	"time"
 
 	log "github.com/sirupsen/logrus"
 
-	"github.com/plunder-app/plunder/pkg/parlay/plugin"
+	parlayplugin "github.com/plunder-app/plunder/pkg/parlay/plugin"
 	"github.com/plunder-app/plunder/pkg/parlay/types"
 
 	"github.com/plunder-app/plunder/pkg/ssh"
 )
-
-var restore Restore
-
-func (m *TreasureMap) findDeployments(deployment []string) (*TreasureMap, error) {
-
-	var newDeploymentList []Deployment
-
-	for x := range deployment {
-		for y := range m.Deployments {
-			if m.Deployments[y].Name == deployment[x] {
-				newDeploymentList = append(newDeploymentList, m.Deployments[y])
-			}
-		}
-	}
-	// If this is zero it means that no deployments have been found
-	if len(m.Deployments) == 0 {
-		return nil, fmt.Errorf("No Deployment(s) have been found")
-	}
-	m.Deployments = newDeploymentList
-	return m, nil
-}
-
-func (d *Deployment) findHosts(hosts []string) (*Deployment, error) {
-
-	var newHostList []string
-
-	for x := range hosts {
-		for y := range d.Hosts {
-			if d.Hosts[y] == hosts[x] {
-				newHostList = append(newHostList, d.Hosts[y])
-			}
-		}
-	}
-	// If this is zero it means that no hosts have been found
-	if len(d.Hosts) == 0 {
-		return nil, fmt.Errorf("No Host(s) have been found")
-	}
-	d.Hosts = newHostList
-	return d, nil
-}
-
-func (d *Deployment) findActions(actions []string) ([]types.Action, error) {
-	var newActionList []types.Action
-
-	for x := range actions {
-		for y := range d.Actions {
-			if d.Actions[y].Name == actions[x] {
-				newActionList = append(newActionList, d.Actions[y])
-			}
-		}
-	}
-	// If this is zero it means that no hosts have been found
-	if len(d.Actions) == 0 {
-		return nil, fmt.Errorf("No Action(s) have been found")
-	}
-	return newActionList, nil
-}
-
-//FindDeployment - takes a number of flags and builds a new map to be processed
-func (m *TreasureMap) FindDeployment(deployment, action, host, logFile string, resume bool) error {
-	var foundMap TreasureMap
-	if deployment != "" {
-		log.Debugf("Looking for deployment [%s]", deployment)
-		for x := range m.Deployments {
-			if m.Deployments[x].Name == deployment {
-				foundMap.Deployments = append(foundMap.Deployments, m.Deployments[x])
-				// Find a specific action and add or resume from
-				if action != "" {
-					// Clear the slice as we will be possibly adding different actions
-					foundMap.Deployments[0].Actions = nil
-					for y := range m.Deployments[x].Actions {
-						if m.Deployments[x].Actions[y].Name == action {
-							// If we're not resuming that just add the action that we want to happen
-							if resume != true {
-								foundMap.Deployments[0].Actions = append(foundMap.Deployments[0].Actions, m.Deployments[x].Actions[y])
-							} else {
-								// Alternatively add all actions from this point
-								foundMap.Deployments[0].Actions = m.Deployments[x].Actions[y:]
-							}
-						}
-					}
-					// If this is zero it means that no actions have been found
-					if len(foundMap.Deployments[0].Actions) == 0 {
-						return fmt.Errorf("No actions have been found, looking for action [%s]", action)
-					}
-				}
-				// If a host is specified act soley on it
-				if host != "" {
-					// Clear the slice as we will be possibly adding different actions
-					foundMap.Deployments[0].Hosts = nil
-					for y := range m.Deployments[x].Hosts {
-
-						if m.Deployments[x].Hosts[y] == host {
-							foundMap.Deployments[0].Hosts = append(foundMap.Deployments[0].Hosts, m.Deployments[x].Hosts[y])
-						}
-					}
-					// If this is zero it means that no hosts have been found
-					if len(foundMap.Deployments[0].Hosts) == 0 {
-						return fmt.Errorf("No host has been found, looking for host [%s]", host)
-					}
-				}
-			}
-		}
-		// If this is zero it means that no actions have been found
-		if len(foundMap.Deployments) == 0 {
-			return fmt.Errorf("No deployment has been found, looking for deployment [%s]", deployment)
-		}
-	} else {
-		return fmt.Errorf("No deployment was specified")
-	}
-	return foundMap.DeploySSH(logFile)
-}
 
 // DeploySSH - will iterate through a deployment and perform the relevant actions
 func (m *TreasureMap) DeploySSH(logFile string) error {
@@ -356,93 +241,4 @@ func parallelDeployment(action []types.Action, hosts []ssh.HostSSHConfig) error 
 		}
 	}
 	return nil
-}
-
-func buildCommand(a types.Action) (string, error) {
-	var command string
-
-	// An executable Key takes presedence
-	if a.KeyName != "" {
-		keycmd := Keys[a.KeyName]
-		// Check that the key exists
-		if keycmd == "" {
-			return "", fmt.Errorf("Unable to find command under key '%s'", a.KeyName)
-
-		}
-		if a.CommandSudo != "" {
-			// Add sudo to the Key command
-			command = fmt.Sprintf("sudo -n -u %s %s", a.CommandSudo, keycmd)
-		} else {
-			command = keycmd
-		}
-	} else {
-		// Not using a key, using a shell command
-		if a.CommandSudo != "" {
-			// Add sudo to the Shell command
-			command = fmt.Sprintf("sudo -n -u %s %s", a.CommandSudo, a.Command)
-		} else {
-			command = a.Command
-		}
-	}
-	return command, nil
-}
-
-func parseAndExecute(a types.Action, h *ssh.HostSSHConfig) ssh.CommandResult {
-	// This will parse the options passed in the action and execute the required string
-	var cr ssh.CommandResult
-	var b []byte
-
-	command, err := buildCommand(a)
-	if err != nil {
-		cr.Error = err
-		return cr
-	}
-
-	if a.CommandLocal == true {
-		log.Debugf("Command [%s]", command)
-		cmd := exec.Command("bash", "-c", command)
-		b, cr.Error = cmd.CombinedOutput()
-		if cr.Error != nil {
-			return cr
-		}
-		cr.Result = strings.TrimRight(string(b), "\r\n")
-	} else {
-		log.Debugf("Executing command [%s] on host [%s]", command, h.Host)
-		cr = ssh.SingleExecute(command, a.CommandPipeFile, a.CommandPipeCmd, *h, a.Timeout)
-
-		cr.Result = strings.TrimRight(cr.Result, "\r\n")
-
-		// If the command hasn't returned anything, put a filler in
-		if cr.Result == "" {
-			cr.Result = "[No Output]"
-		}
-		if cr.Error != nil {
-			return cr
-		}
-	}
-
-	// Save the results into a key to be used at another point
-	if a.CommandSaveAsKey != "" {
-		log.Debugf("Adding new results to key [%s]", a.CommandSaveAsKey)
-		Keys[a.CommandSaveAsKey] = cr.Result
-	}
-
-	// Save the results into a file to be used at another point
-	if a.CommandSaveFile != "" {
-		var f *os.File
-		f, cr.Error = os.Create(a.CommandSaveFile)
-		if cr.Error != nil {
-			return cr
-		}
-
-		defer f.Close()
-
-		_, cr.Error = f.WriteString(cr.Result)
-		if cr.Error != nil {
-			return cr
-		}
-		f.Sync()
-	}
-
-	return cr
 }
